1.7 changes to script.js and promise-polyfill.js inclusion
Removed static pokemon repository, added pokemonList  as empty array, and pointed data to a pokemon api limiting to the first 151 pokemon.

Reused the add pokemon function, but added an if/else statement that looks for a name element, otherwise lists an error.

Removed function test that added a pokemon after the fact using the add function.

Created function loadList that fetches the pokemon api and converts it to json. The json output is then looped and pokemon is defined name and details. This is put into the add function to display on the page as li elements.

Function loadDetails is added that uses the pokemon api and returns a json converted format from the pokemonList detailsUrl location. Details pulled are an image, height, and types.

Event listener now points to loadDetails funciton and logs to console.